[Refactor]: 디렉토리 추가
